Add Mouse Events and Supporting Models to Vue Demos
The update introduces mouse events handling in the Vue3 examples. New demo tests and models were added for this functionality, including extending the AreaModel for Rowspan and Colspan. Also, the "mouseevents" demo was added to the application router. The demo features include context menu, click and move events, with corresponding models for data manipulation. In the angular table demo, unnecessary imports were removed.