Update blackout demo and shift styles to separate file
This commit refactors the structure for the blackout demo in the GUI expert app. The CSS styles that were initially part of the run.astro file are now held in a separate blackout.css style file. This improves code structure and organization for better manageability. Additionally, the class name has been updated to 'blackout-demo' for better context.